backend/requirements.txt
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-jose==3.3.0
python-multipart==0.0.9